requirements.txt
Django==2.1
django-extensions==2.1.2
django-modern-rpc==0.11.1
pytz==2018.5
six==1.11.0
